+++ firestore-bigquery-export/functions/src/index.ts
@@ -29,6 +29,7 @@ const eventTracker: FirestoreEventHistoryTracker = new FirestoreBigQueryEventHis
     tableId: config.tableId,
     datasetId: config.datasetId,
     datasetLocation: config.datasetLocation,
+    tablePartitioning: config.tablePartitioning,
   }
 );